Ensure 'flush' is called in workers
In batch sending mode reports are usually sent on shutdown but workers
are (generally) long running processes so this doesn't work. Instead we
flush after each handled message and on failed messages

If we're not in batch sending mode then there could still be reports
to send, e.g. if batch sending was toggled part way through handling
a message. This is why we call 'flush' unconditionally. If there is
nothing to send then 'flush' is a noop anyway