Use build args to inject environment variables
Previously the environment variables were not available until runtime.
This is problematic because Symfony builds the DI container after
running composer install. If the build requires an env var, such as
when Bugsnag is used (with our app's configuration), then this build
won't work because the required env var won't be present

We don't need these variables to be modifiable at runtime - this was
only the case because this setup was copied from laravel MR tests -
so we can safely use build args here instead