Setup bugsnag.yaml at build time
This allows us to test that config values in the YAML config are
used, which is better than the previous system of using query string
parameters

We can't use env variables for this at runtime as certain types of
config values cannot be dynamic, e.g. arrays for 'redacted_keys'

This is a restriction Symfony places on config values so that it can
validate the correct type is always used. Using a string from env
is fine because env vars are always strings